refactor(schema-api): extract SecurityApi trait
Extract SecurityApi trait from monolithic SchemaApi to improve code
organization and maintainability. This continues the ongoing schema API
refactoring effort by moving security policy management operations into
their own focused trait interface.

The extraction includes security policy operations:
- set_table_column_mask_policy: Set column data masking policies for tables
- set_table_row_access_policy: Set row-level access policies for tables

Changes made:
- Created new `security_api.rs` with SecurityApi trait definition
- Moved 2 security methods with complete implementations from SchemaApi
- Moved helper function `update_mask_policy` for policy management
- Updated SchemaApi trait bounds to include SecurityApi dependency
- Added module exports and fixed import dependencies in consuming code
- Updated mutable_catalog.rs and default_database.rs to import SecurityApi

Progress: 55 of 58 methods (95%) now extracted from SchemaApi into
focused sub-traits, leaving only minimal specialized domains for final
cleanup: GC (1), Utility (2) methods.
