+++ src/meta/api/src/lib.rs
@@ -35,6 +35,7 @@ pub mod name_value_api;
 pub mod reply;
 mod schema_api;
 mod schema_api_test_suite;
+pub mod security_api;
 mod sequence_api;
 pub mod table_api;
 pub(crate) mod testing;
@@ -57,6 +58,7 @@ pub use lock_api::LockApi;
 pub use row_access_policy_api::RowAccessPolicyApi;
 pub use schema_api::SchemaApi;
 pub use schema_api_test_suite::SchemaApiTestSuite;
+pub use security_api::SecurityApi;
 pub use sequence_api::SequenceApi;
 pub use table_api::TableApi;
 pub use util::assert_table_exist;
