+++ src/query/service/src/catalogs/default/mutable_catalog.rs
@@ -30,6 +30,7 @@ use databend_common_meta_api::DictionaryApi;
 use databend_common_meta_api::IndexApi;
 use databend_common_meta_api::LockApi;
 use databend_common_meta_api::SchemaApi;
+use databend_common_meta_api::SecurityApi;
 use databend_common_meta_api::SequenceApi;
 use databend_common_meta_api::TableApi;
 use databend_common_meta_app::app_error::AppError;
