+++ src/query/service/src/databases/default/default_database.rs
@@ -16,7 +16,7 @@ use std::sync::Arc;
 
 use databend_common_catalog::table::Table;
 use databend_common_exception::Result;
-use databend_common_meta_api::SchemaApi;
+use databend_common_meta_api::SecurityApi;
 use databend_common_meta_api::TableApi;
 use databend_common_meta_app::app_error::AppError;
 use databend_common_meta_app::app_error::UnknownTable;
